drift-service/requirements.txt
redis==5.0.1
numpy==1.26.3
scipy==1.13.0
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
psycopg2-binary==2.9.9